Create signal receiver to update the number of likes once it is updated
You have to connect your receiver function to a signal so that it gets called every time the signal is sent. The recommended method for registering your signals is
by importing them in the ready() method of your application configuration class. Django provides an application registry that allows you to configure and introspect your applications.